backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.32.5
PyJWT==2.11.0
cryptography==46.0.5
pypdf2==3.0.1